server/requirements.txt
Flask
gunicorn
flask-cors
python-dotenv
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client